Joli Makefile, joli Maaaaakefile

+++ Makefile
@@ -1,8 +1,11 @@
 all:
-	ocamlbuild -use-menhir main.native -j 4
+	ocamlbuild -use-menhir -menhir "menhir -v" main.native -j 4
+	@echo
+	@echo "Moving binary to" `pwd`
 	mv _build/main.native minic++
 	rm main.native
 
 clean:
 	ocamlbuild -clean
-	rm minic++
+	@echo "Removing binary : rm minic++"
+	if [ -e minic++ ]; then rm minic++; fi
